Added the energy condenser
See a minecraft wiki for explanation of what the energy condenser does. This version is currently using placeholders for the picture and recipe.  There are currently a few known bugs with netcode, and with the energy condenser not closing automatically when the player moves too far away. There is also a lack of major quality of life features such as a quick loot button to quickly loot the condenser.