Several QOL and documentation features
Added a loot all button to the energy condenser which will quickly loot all items out of the condenser, while leaving its template unaffected

Added a custom mod icon

Added a search box to the transmutation tablet for looking up items with names that have a substring matching the provided string in the search box. Incidentally, created a new UI element for drawing text to the screen.

Updated the description and added a workshop description