Numerous changes related to tablet UI
Added the ability to instantly transmute the held item if the transmutation tablet UI is open and the shift key is pressed.

Fixed an issue where multiple versions of the shellphone could be learned.

Updated the transmutation tablet UI's EMC display to show EMC as a double to 6 decimal places if hovered over, and also changed it to look more like a fraction displayed in LaTeX.

Added a new UI element for displaying fractions.